+++ app.js
@@ -805,8 +805,9 @@ spacialistApp.filter('urlify', function() {
     };
 });
 
-spacialistApp.filter('csv2table', function() {
-    return function(csv, delimiter, rowCount) {
+spacialistApp.filter('csv2table', function($translate) {
+    return function(csv, delimiter, hasHeader, rowCount) {
+        if(hasHeader !== false) hasHeader = true;
         if(!csv) return '<table></table>';
         delimiter = delimiter || ',';
         var rendered = '<table class="table table-striped">';
@@ -814,9 +815,15 @@ spacialistApp.filter('csv2table', function() {
             delimiter = '\t';
         }
         var dsv = d3.dsv(delimiter);
+        var rows;
         var headers = csv.split('\n')[0];
         var headerCols = dsv.parseRows(headers)[0];
-        var rows = dsv.parse(csv);
+        if(hasHeader) {
+            rows = dsv.parse(csv);
+        } else {
+            headerCols = createCountingCsvHeader(headerCols.length, $translate);
+            rows = dsv.parseRows(csv);
+        }
 
         rendered += '<tr>';
         for(var i=0; i<headerCols.length; i++) {